Added Title and BG
I made a custom title for the homepage, I'll prolly have to change it to work for the top bar but I'll do that when I get to it. I also was able to make the background be a bunch of moving stars, I want to see if I can change the speed and direction but again something for later.